extract syringe api to separated function
