+++ service-jehlomat/src/main/kotlin/main/Application.kt
@@ -1,5 +1,6 @@
 package main
 
+import api.syringeApi
 import api.googlePlacesProxy
 import api.organizationApi
 import api.teamApi
@@ -72,56 +73,8 @@ fun Application.module(testing: Boolean = false) {
     )
 
     routing {
-        route("/api/v1/jehlomat") {
-
-            get("/syringes") {
-                val format = SimpleDateFormat.getInstance()
-                val parameters = call.request.queryParameters
-
-                val city = parameters["city"]
-                val username = parameters["username"]
-                val from = Date(parameters["from"]?.toLong() ?: 0L)
-                val to = Date(parameters["to"]?.toLong() ?: System.currentTimeMillis())
-                val liquidationStatus = try {
-                    parameters["status"]?.let { Demolisher.valueOf(it) } ?: run { Demolisher.NO }
-                } catch (ex: IllegalArgumentException) {
-                    Demolisher.NO
-                }
-
-                // pass to database query and retrieve from it
-
-                call.respond(listOf(syringe))
-            }
-
-            route("/syringe/{id}") {
-                get {
-                    val id = call.parameters["id"]?.toLong()
-                    try {
-                        val filteredSyringe = syringes.filter { it.id == id }[0]
-                        call.respond(HttpStatusCode.OK, filteredSyringe)
-                    } catch (ex: IndexOutOfBoundsException) {
-                        call.respond(HttpStatusCode.NotFound)
-                    }
-                }
-
-                put {
-                    val id = call.parameters["id"]?.toLong()
-                    syringes.removeIf { it.id == id }
-                    syringes.add(call.receive())
-                    call.respond(HttpStatusCode.OK)
-                }
-
-                delete {
-                    val id = call.parameters["id"]?.toLong()
-                    syringes.removeIf { it.id == id }
-                    call.respond(HttpStatusCode.OK)
-                }
-            }
-
-            post("/syringe") {
-                syringes.add(call.receive())
-                call.respond(HttpStatusCode.Created)
-            }
+        route("/api/v1/jehlomat/syringe") {
+            syringeApi()
         }
         route("/api/v1/jehlomat/users") {
             userApi()
